# pims/pims/openapi/api-specification.yaml
openapi: 3.0.3
info:
  title: Cytomine Python Image Managament Server PIMS
  description: |
    Cytomine Python Image Managament Server (PIMS) HTTP API. While this API is intended to be internal, a lot of the
    following specification can be ported to the external Cytomine API.

    # Additional API formats
    In addition to OpenAPI types and formats, the following table contains some additional formats used in this
    specification.

    | Type | Format | Examples | Specification |
    |------|--------|----------|---------------|
    | `string` | `range` | **"1:3"**, **"2:"**, **":"**| Inspired from Python slicing |
    | `string` | `wkt` | **"POINT (10 20)"** | [Open Geospacial Consortium](https://www.ogc.org/standards/sfa)|
    | `string` | `color` | **"#FF00AA"**, **"rgb(100,0,0)"**, **"red"** | [CSS Color Module Level 3](https://drafts.csswg.org/css-color-3/#colorunits)|
    | `string` | `colormap-id` | **"jet"**, **"!jet"**||
    | `string` | `filter-id` | **"yen"** ||
  version: 0.1.0
#servers:
#  - url: 'https'
...
paths:
...
  /image/{filepath}/window:
    post:
      operationId: pims.api.window.show_window_with_body
      summary: Get a window
      description: |
        **`GET with body` - when a GET with URL encoded query parameters is not possible due to URL size limits, a POST
        with body content must be used.**

        Get a window (rectangular crop) extract from an image, with given channels, focal
        planes and timepoints. If multiple channels are given (slice or selection), they are merged. If multiple focal
        planes or timepoints are given (slice or selection), a reduction function must be provided.

        **By default**, all image channels are used and when the image is multidimensional, the
         tile is extracted from the median focal plane at first timepoint.
      tags:
        - Windows
      parameters:
        - $ref: '#/components/parameters/path-filepath'
        - $ref: '#/components/parameters/header-etag'
        - $ref: '#/components/parameters/header-x-image-size-safety'
        - $ref: '#/components/parameters/header-x-annotation-origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - title: input-region-target-length
                  allOf:
                    - $ref: '#/components/schemas/region-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-length-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-region-target-width
                  allOf:
                    - $ref: '#/components/schemas/region-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-width-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-region-target-height
                  allOf:
                    - $ref: '#/components/schemas/region-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-height-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-region-target-zoom
                  allOf:
                    - $ref: '#/components/schemas/region-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-zoom-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-region-target-level
                  allOf:
                    - $ref: '#/components/schemas/region-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-level-object-required'
                    - $ref: '#/components/schemas/window-basis'


                - title: input-tile-index-target-length
                  allOf:
                    - $ref: '#/components/schemas/tile-index-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-length-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-tile-index-target-width
                  allOf:
                    - $ref: '#/components/schemas/tile-index-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-width-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-tile-index-target-height
                  allOf:
                    - $ref: '#/components/schemas/tile-index-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-height-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-tile-index-target-zoom
                  allOf:
                    - $ref: '#/components/schemas/tile-index-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-zoom-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-tile-index-target-level
                  allOf:
                    - $ref: '#/components/schemas/tile-index-object'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-level-object-required'
                    - $ref: '#/components/schemas/window-basis'

                - title: input-tile-coord-target-length
                  allOf:
                    - $ref: '#/components/schemas/tile-coord'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-length-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-tile-coord-target-width
                  allOf:
                    - $ref: '#/components/schemas/tile-coord'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-width-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-tile-coord-target-height
                  allOf:
                    - $ref: '#/components/schemas/tile-coord'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-height-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-tile-coord-target-zoom
                  allOf:
                    - $ref: '#/components/schemas/tile-coord'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-zoom-object-required'
                    - $ref: '#/components/schemas/window-basis'
                - title: input-tile-coord-target-level
                  allOf:
                    - $ref: '#/components/schemas/tile-coord'
                    - $ref: '#/components/schemas/tier-index'
                    - $ref: '#/components/schemas/target-level-object-required'
                    - $ref: '#/components/schemas/window-basis'
      responses:
        200:
          $ref: '#/components/responses/window'
        304:
          $ref: '#/components/responses/not-modified'
        400:
          $ref: '#/components/responses/problem-bad-request'
        404:
          $ref: '#/components/responses/problem-not-found'
        406:
          $ref: '#/components/responses/problem-not-acceptable'
...
components:
  parameters:
    header-etag:
      name: ETag
      in: header
      description: |
        The RFC 7232 ETag header field in a response provides the entity-tag of
        a selected resource. The entity-tag is an opaque identifier for versions
        and representations of the same resource over time, regardless whether
        multiple versions are valid at the same time. An entity-tag consists of
        an opaque quoted string, possibly prefixed by a weakness indicator (see
        [RFC 7232 Section 2.3](https://tools.ietf.org/html/rfc7232#section-2.3)).
      required: false
      schema:
        type: string
    header-x-image-size-safety:
      name: X-Image-Size-Safety
      in: header
      description: |
        This header provides hints about the way the server has to deal with too large image responses.
        * `SAFE_REJECT` - Reject too large image response and throw a `400 Bad Request`.
        * `SAFE_RESIZE` - Resize the image response to an acceptable image size. Information about the resize are
        returned in `X-Image-Size-Limit` header.
        * `UNSAFE` - **At your own risk!** Try to fulfill the request but can cause unintended side effects
        (unreadable response, server slown down, server failure, ...). It should only be used in rare controlled
        situations.
      required: false
      schema:
        default: SAFE_RESIZE
        type: string
        enum:
          - SAFE_REJECT
          - SAFE_RESIZE
          - UNSAFE
    header-x-annotation-origin:
      name: X-Annotation-Origin
      in: header
      description: |
        This header give the origin coordinate system in which are described provided annotations.
      required: false
      schema:
        default: LEFT_TOP
        type: string
        enum:
          - LEFT_TOP
          - LEFT_BOTTOM
    path-filepath:
      name: filepath
      in: path
      description: The file path, relative to server base path.
      required: true
      schema:
        type: string
        format: path
      example: '123/my-file.ext'
...
  schemas:
...
    single-channel-index:
      type: integer
      format: int32
      description: A single channel index. **By default**, all channels are considered.
      minimum: 0
...
    window-plane-selection:
      type: object
      properties:
        channels:
          $ref: '#/components/schemas/single-channel-index'
        z_slices:
          $ref: '#/components/schemas/single-z-slice-index'
        timepoints:
          $ref: '#/components/schemas/single-timepoint-index'
#        c_reduction:
#          $ref: '#/components/schemas/channel-reduction'
#        z_reduction:
#          $ref: '#/components/schemas/z-slice-reduction'
#        t_reduction:
#          $ref: '#/components/schemas/timepoint-reduction'
...
    zoom-or-level:
      type: integer
      format: int64
      minimum: 0
      example: 0
    tile-index:
      description: |
        The tile index is the position of the tile in the given image pyramid tier.
        It is computed as `n * n_x_tiles + m` where
        * `n_x_tiles` is the number of tiles along the horizontal axis at given tier.
        * `m` is the tile position along the horizontal axis at given tier (0 is left).
        * `n` is the tile position along the vertical axis at given tier (0 is top).
      type: integer
      format: int64
      minimum: 0
    tile-x:
      description: |
         The tile position along the horizontal axis at given tier (0 is left).
      type: integer
      format: int64
      minimum: 0
    tile-y:
      description: |
        The tile position along the vertical axis at given tier (0 is top).
      type: integer
      format: int64
      minimum: 0
    region:
      type: object
      description: The input region in the reference tier.
      required:
        - top
        - left
        - width
        - height
      properties:
        top:
          allOf:
            - description: |
                The input region top coordinate in the reference tier. The position `0` denotes the
                top-most pixels in the image.
            - $ref: '#/components/schemas/region-coordinate'
        left:
          allOf:
            - description: |
                The input region left coordinate in the reference tier. The position `0` denotes the left-most pixels
                in the image.
            - $ref: '#/components/schemas/region-coordinate'
        width:
          allOf:
            - description: |
                The input region width in the reference tier.
            - $ref: '#/components/schemas/region-length'
        height:
          allOf:
            - description: |
                The input region height in the reference tier.
            - $ref: '#/components/schemas/region-length'
    region-coordinate:
      type: integer
      format: int64
      description: A coordinate in pixels in the reference tier.
      minimum: 0
      example: 256
#        - type: number
#          format: double
#          description: A coordinate expressed in percentage relatively to the reference tier.
#          minimum: 0.0
#          maximum: 1.0
#          example: 0.5
    region-length:
      type: integer
      format: int64
      description: A length in pixels in the reference tier.
      minimum: 1
      example: 256
#        - type: number
#          format: double
#          description: A length expressed in percentage relatively to the reference tier.
#          minimum: 0.0
#          maximum: 1.0
#          example: 0.5

    # Output sizes
    target-length:
      description: |
        Length of the largest side of the target. The other dimension is adjusted to preserve the aspect
        ratio.
      example: 256
      oneOf:
        - type: integer
          format: int64
          description: A length in pixels.
          minimum: 1
          example: 256
        - type: number
          format: double
          description: A length expressed in percentage relatively to the real image length.
          minimum: 0.0
          maximum: 1.0
          example: 0.5
...
    target-length-object-required:
      type: object
      required:
        - length
      properties:
        length:
          $ref: '#/components/schemas/target-length'
...
    target-width:
      description: Width of the target. Height is adjusted to preserve the aspect ratio.
      allOf:
        - $ref: '#/components/schemas/target-length'
    target-width-object-required:
      type: object
      required:
        - width
      properties:
        width:
          $ref: '#/components/schemas/target-width'
...
    target-height:
      description: Height of the target. Width is adjusted to preserve the aspect ratio.
      allOf:
        - $ref: '#/components/schemas/target-length'
    target-height-object-required:
      type: object
      required:
        - height
      properties:
        height:
          $ref: '#/components/schemas/target-height'
...
    target-zoom:
      description: |
        The zoom level to consider for the target.
        Zoom 0 has the worst resolution (smallest image, top of the image pyramid).
        Maximum admissible zoom level depends on image.
      allOf:
        - $ref: '#/components/schemas/zoom-or-level'
    target-zoom-object-required:
      type: object
      required:
        - zoom
      properties:
        zoom:
          $ref: '#/components/schemas/target-zoom'
...
    target-level:
      description: |
        The tier level to consider for the target.
        Level 0 has the best resolution (largest image, basis of the image pyramid).
        Maximum admissible tier level depends on image.     
      allOf:
        - $ref: '#/components/schemas/zoom-or-level'
    target-level-object-required:
      type: object
      required:
        - level
      properties:
        level:
          $ref: '#/components/schemas/target-level'
...
    tile-coord:
      type: object
      required:
        - tx
        - ty
      properties:
        tx:
          $ref: '#/components/schemas/tile-x'
        ty:
          $ref: '#/components/schemas/tile-y'
    tile-index-object:
      type: object
      required:
        - ti
      properties:
        ti:
          $ref: '#/components/schemas/tile-index'
    tier-index:
      type: object
      properties:
        reference_tier_index:
          type: integer
          description: |
            The index (zoom or level) denoting the tier to use as reference coordinate system for the input region. By
            default, best resolution tier (maximum zoom/lowest level) is used.
        tier_index_type:
          type: string
          enum:
            - LEVEL
            - ZOOM
          default: LEVEL
          description: |
            How to interprete `reference_tier_index`, either as a zoom index, either as a level index.
    region-object:
      type: object
      required:
        - region
      properties:
        region:
          $ref: '#/components/schemas/region'
...
    resized-basis:
      allOf:
        - $ref: '#/components/schemas/window-plane-selection'
        - $ref: '#/components/schemas/image-manipulation'
        - type: object
          properties:
            bits:
              $ref: '#/components/schemas/bits'
            colorspace:
              $ref: '#/components/schemas/colorspace'
    window-basis:
      allOf:
        - $ref: '#/components/schemas/resized-basis'
        - type: object
          properties:
            annotations:
              $ref: '#/components/schemas/annotations'
            annotation_style:
              oneOf:
                - title: MASK
                  type: object
                  required:
                    - mode
                  properties:
                    mode:
                      description: The mode to use to render annotations
                      type: string
                      enum:
                        - MASK
                - title: CROP
                  type: object
                  required:
                    - mode
                  properties:
                    mode:
                      description: The mode to use to render annotations
                      type: string
                      enum:
                        - CROP
                    background_transparency:
                      $ref: '#/components/schemas/annotation-bg-transparency'
                - title: DRAWING
                  allOf:
                    - type: object
                      required:
                        - mode
                      properties:
                        mode:
                          description: The mode to use to render annotations
                          type: string
                          enum:
                            - DRAWING
                    - $ref: '#/components/schemas/annotation-point-drawing-object'
    # Annotation crops
...
    annotations:
      x-single-or-array: true
      oneOf:
        - $ref: '#/components/schemas/annotation'
        - type: array
          items:
            $ref: '#/components/schemas/annotation'
...
    annotation-bg-transparency:
      type: integer
      format: int32
      minimum: 0
      maximum: 100
      default: 100
      description: |
        The background transparency. 100 means transparent background. When transparency is used, the target
        content type must be an image format supporting transparency.
...
    annotation-point-drawing-object:
      type: object
      properties:
        point_envelope_length:
          type: integer
          default: 100
          description: |
            Point geometries have no area. An envelope must be specified to extract some image data.
            This envelope is a square of length given by this parameter, whose the center is the given point.
            This parameter has no effect on non-point geometries.
        point_cross:
          type: string
          enum:
            - CROSS
            - CROSSHAIR
            - CIRCLE
          default: CROSS
          description: |
            Point geometries have no area. Possible representation as a drawing are:
            * `CROSS` - A regular cross, hiding point location
            * `CROSSHAIR` - A cross but whose intersection is removed to show the point locatio
            * `CIRCLE` - A circle around the point location
            This parameter has no effect on non-point geometries.
...
    generic-image:
      type: string
      format: binary
      example: (binary)
    problem:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: |
            An URI that identifies the problem type.  When dereferenced,
            it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          default: 'about:blank'
          example: 'problem/constraint-violation'
        title:
          type: string
          description: |
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized).
          example: Service Unavailable
        status:
          type: integer
          format: int32
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          minimum: 400
          maximum: 600
          exclusiveMaximum: true
          example: 503
        detail:
          type: string
          description: |
            A human readable explanation specific to this occurrence of the
            problem.
          example: Connection to database timed out
        instance:
          type: string
          format: uri
          description: |
            An absolute URI that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced.
    invalid-parameters-problem:
      description: Problem details for invalid input parameter(s)
      allOf:
        - $ref: '#/components/schemas/problem'
        - type: object
          properties:
            invalid-parameters:
              type: array
              description: An array of parameter violations
              items:
                $ref: '#/components/schemas/invalid-parameter'
...
  headers:
    If-None-Match:
      description: |
        The RFC7232 If-None-Match header field in a request requires the server
        to only operate on the resource if it does not match any of the provided
        entity-tags. If the provided entity-tag is `*`, it is required that the
        resource does not exist at all (see [RFC 7232 Section
        3.2](https://tools.ietf.org/html/rfc7232#section-3.2)).
      required: false
      schema:
        type: string
    Cache-Control:
      description: |
        The RFC 7234 Cache-Control header field is providing directives to
        control how proxies and clients are allowed to cache responses results
        for performance. Clients and proxies are free to not support caching of
        results, however if they do, they must obey all directives mentioned in
        [RFC-7234 Section 5.2.2](https://tools.ietf.org/html/rfc7234) to the
        word.

        In case of caching, the directive provides the scope of the cache
        entry, i.e. only for the original user (private) or shared between all
        users (public), the lifetime of the cache entry in seconds (max-age),
        and the strategy how to handle a stale cache entry (must-revalidate).
        Please note, that the lifetime and validation directives for shared
        caches are different (s-maxage, proxy-revalidate).
      required: false
      schema:
        type: string
        example: "private, must-revalidate, max-age=300"
    Vary:
      description: |
        The RFC 7231 Vary header field in a response defines which parts of
        a request message, aside the target URL and HTTP method, might have
        influenced the response. A client or proxy cache must respect this
        information, to ensure that it delivers the correct cache entry (see
        [RFC-7231 Section
        7.1.4](https://tools.ietf.org/html/rfc7231#section-7.1.4)).
      required: false
      schema:
        type: string
        example: "accept-encoding, accept-language"
...
    X-Image-Size-Limit:
      description: |
        This header indicates that the image representation has been resized due to reached image size limit. This
        limit can be set by the server or by the format itself.
      required: false
      style: simple
      explode: true
      schema:
        type: object
        properties:
          ratio:
            type: number
          request_width:
            type: integer
          request_height:
            type: integer
          safe_width:
            type: integer
          safe_height:
            type: integer
  responses:
    problem-bad-request:
      description: Bad request - Input payload fails business logic validation or generic error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/invalid-parameters-problem'
    problem-not-found:
      description: Not Found - The resource is not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/problem'
    problem-not-acceptable:
      description: |
        Not Acceptable - The resource can only generate content not acceptable according to the Accept
        headers sent in the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/problem'
...
    not-modified:
      description: Not Modified
...
    window:
      description: OK - The window
      headers:
        If-None-Match:
          $ref: '#/components/headers/If-None-Match'
        Cache-Control:
          $ref: '#/components/headers/Cache-Control'
        Vary:
          $ref: '#/components/headers/Vary'
        X-Image-Size-Limit:
          $ref: '#/components/headers/X-Image-Size-Limit'
      content:
        image/jpeg:
          schema:
            $ref: '#/components/schemas/generic-image'
        image/png:
          schema:
            $ref: '#/components/schemas/generic-image'
        image/webp:
          schema:
            $ref: '#/components/schemas/generic-image'
...